Refactor: Moves Scaffold to OwlApp file
